requirements.txt
instagrapi>=2.0.0
moviepy>=1.0.3
Pillow>=10.0.0
